Use separate build and install scripts in travis
This approach has several advantages:

(1) If any step in the build fails the follow steps
are not run.  With travis the scripts are run
consecutively even if an earlier one failed.

(2) Makes it easier to run the scripts locally
outside of travis CI, for debugging.

(3) The syntax is a lot clearer.

Also, switch to container-based travis runs by
setting sudo to false. This allows for faster
startup.

Also, enable testing of the NaCl build in travis
now, since the issue that previously prevented
this has now been fixed